backend/requirements.txt
fastapi
uvicorn
google-generativeai
requests
beautifulsoup4
python-multipart
python-dotenv
